+++ CHANGELOG.md
@@ -6,11 +6,13 @@
 - Fixed a bug that did not update the modification time of a file on save.
 - Fixed the non-selection of the current file on directory selection
 - Fixed a small bug that sometimes could throw an error when moving directories within the app
+- Now a quicklook window can be brought to front by simply clicking its title
 
 ## Under the hood
 
 - Begun another code rewrite. This time, the focus is on two parts: First, prevent any access of object properties from other objects than `this`, instead, use public functions (also, prevent calling of private functions from the outside). Second: Try to, again, move out some functionality from the Zettlr main function to shorten it.
 - Removed unnecessary function calls in the renderer.
+- Also added support for the [yarn package manager](https://yarnpkg.com/)
 
 # 0.13.0